Migrate to clrmd 2

+++ JitDasm/MethodJitter.cs
@@ -68,7 +68,7 @@ namespace JitDasm {
 				}
 			}
 			foreach (var type in allTypes) {
-				if (!typeFilter.IsMatch(MakeClrmdTypeName(type.FullName ?? string.Empty), (uint)type.MetadataToken))
+				if (!typeFilter.IsMatch(MakeClrmdTypeName(type.FullName ?? string.Empty), type.MetadataToken))
 					continue;
 				bool isDelegate = typeof(Delegate).IsAssignableFrom(type);
 				foreach (var method in GetMethods(type)) {
@@ -76,7 +76,7 @@ namespace JitDasm {
 						continue;
 					if (method.IsGenericMethod)
 						continue;
-					if (!methodFilter.IsMatch(method.Name, (uint)method.MetadataToken))
+					if (!methodFilter.IsMatch(method.Name, method.MetadataToken))
 						continue;
 					// Not supported on .NET Core
 					if (isDelegate && method is MethodInfo m && m.IsVirtual && (m.Name == "BeginInvoke" || m.Name == "EndInvoke"))
